move tp3 -> tinkerpop subpackage
