+++ tinkerpop3/src/main/java/overflowdb/OdbGraph.java
@@ -23,12 +23,12 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import overflowdb.storage.NodeDeserializer;
 import overflowdb.storage.OdbStorage;
-import overflowdb.tp3.GraphVariables;
-import overflowdb.tp3.TinkerIoRegistryV1d0;
-import overflowdb.tp3.TinkerIoRegistryV2d0;
-import overflowdb.tp3.TinkerIoRegistryV3d0;
-import overflowdb.tp3.optimizations.CountStrategy;
-import overflowdb.tp3.optimizations.OdbGraphStepStrategy;
+import overflowdb.tinkerpop.GraphVariables;
+import overflowdb.tinkerpop.TinkerIoRegistryV1d0;
+import overflowdb.tinkerpop.TinkerIoRegistryV2d0;
+import overflowdb.tinkerpop.TinkerIoRegistryV3d0;
+import overflowdb.tinkerpop.optimizations.CountStrategy;
+import overflowdb.tinkerpop.optimizations.OdbGraphStepStrategy;
 import overflowdb.util.MultiIterator2;
 import overflowdb.util.NodesList;
 import overflowdb.util.PropertyHelper;
