+++ tinkerpop3/src/main/java/overflowdb/OdbNode.java
@@ -2,7 +2,7 @@ package overflowdb;
 
 import org.apache.commons.lang3.NotImplementedException;
 import org.apache.tinkerpop.gremlin.util.iterator.IteratorUtils;
-import overflowdb.tp3.Converters;
+import overflowdb.tinkerpop.Converters;
 import overflowdb.util.ArrayOffsetIterator;
 import overflowdb.util.MultiIterator2;
 import overflowdb.util.PackedIntArray;
