+++ tinkerpop3/src/main/java/overflowdb/tinkerpop/optimizations/OdbGraphStep.java
@@ -1,4 +1,4 @@
-package overflowdb.tp3.optimizations;
+package overflowdb.tinkerpop.optimizations;
 
 import overflowdb.OdbGraph;
 import org.apache.tinkerpop.gremlin.process.traversal.Compare;
